CLI: Refactoring to avoid namespace clash with public bindings
