+++ Makefile
@@ -3,7 +3,7 @@
 DESTDIR ?= ../tmproot
 
 default: _build
-	meson compile -C _build
+	meson compile -C _build --verbose
 
 _build:
 	meson setup _build --prefix=/usr
@@ -12,7 +12,7 @@ _build-cov:
 	meson setup _build-cov --prefix=/usr -Db_coverage=true
 
 clean:
-	rm -f netplan/_features.py src/_features.h src/_features.h.gch
+	rm -f netplan_cli/_features.py src/_features.h src/_features.h.gch
 	rm -f generate doc/*.html doc/*.[1-9]
 	rm -f *.o *.so*
 	rm -f netplan-dbus dbus/*.service
@@ -35,7 +35,7 @@ linting: _build
 	meson test -C _build --verbose codestyle
 
 pre-coverage: _build-cov
-	meson compile -C _build-cov
+	meson compile -C _build-cov --verbose
 
 check-coverage: pre-coverage
 	meson test -C _build-cov
