+++ netplan_cli/__init__.py
@@ -15,6 +15,6 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from netplan.cli.core import Netplan
+from .cli.core import Netplan
 
 __all__ = [Netplan]
